implement CLI
  - move date/time/datetime sqlite adapters to adapters module
  - move all sqlite adapter logic from __init__ to __main__
  - remove schema option (doesn't make sense for sqlite here)
  - rewrite tests against the CLI (test_cli), disable earlier tests

Note sqlite adapter settings were moved to the CLI vs. the library code,
because the assumption is using the library you bring your own config of
sqlite3 -- you are not limited to the few preset adapters I set up for
date/time/datetime data. Likewise with the config of the sas7bdat parser.
The library assumes you set that up externally. (The one exception being
that it forces skip_header=True, which is always what we want for
importing to a dbase.)